+++ EssentialApp/EssentialAppTests/Helpers/XCTestCase+FeedImageDataLoader.swift
@@ -6,7 +6,7 @@
 //
 
 import XCTest
-import EssentialFeed
+@testable import EssentialFeed
 
 protocol FeedImageDataLoaderTestCase: XCTestCase {}
